Change Python interpreter path in CLI access test
This commit modifies the `test_main_access_cli` test in `test_cli.py` to use the path of the currently active Python interpreter when running the `exif_stripper.cli` module. This change ensures that the test uses the same Python interpreter regardless of where it's run from, which fixes an issue where the test was failing when run from VS Code's testing panel but passing when run from the terminal.

The path of the currently active Python interpreter is obtained using `sys.executable`. This path is then used in the `subprocess.run` command to run the `exif_stripper.cli` module.